% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_output.R
\name{make_output}
\alias{make_output}
\alias{make_output.data.frame}
\alias{make_output.tbl_dbi}
\alias{make_output.connection}
\alias{make_output.default}
\title{Create output sink}
\usage{
make_output(output, ...)

\method{make_output}{data.frame}(output, ...)

\method{make_output}{tbl_dbi}(output, ...)

\method{make_output}{connection}(output, ...)

\method{make_output}{default}(output, ...)
}
\arguments{
\item{output}{an object determining the type of output sink to be created}

\item{...}{named arguments expected to be vectors. Shorter ones are recycled.}
}
\value{
A function, say, \code{output} with signature \code{...} which will write data
supplied in its arguments.
}
\description{
Function \code{\link[=make_output]{make_output()}} returns an output sink function of a type determined
by the class of the \code{output} argument. Output sink is a function that has a
single argument \code{...} through which one can pass multiple named arguments.
These arguments are expected to be vectors of the same length. Repeated calls
to an output sink write supplied data to a target, i.e. append a data frame,
write to a connection, or append a table in a database. Implemented methods
include:

\itemize{
\item Appending an existing data frame
}

\itemize{
\item Appending a table in a database
}

\itemize{
\item Writing to a connection
}

\itemize{
\item Creating and appending a data frame
}
}
\details{
If \code{output} is a data frame then the calls to the output sink
created will append that data frame.

If \code{output} is an object inheriting from \code{tbl_dbi}...

If \code{output} is a connection object then arguments in \code{...} are
passed to \code{\link[=cat]{cat()}} with argument \code{file} given the connection. Argument names
are not included in the output nor checked. The order of the values is
determined by the order of the arguments.

If \code{output} is \code{NULL} then the result is an "blank" data frame
output sink. The first call to the sink will instantiate the columns and
subsequent calls will append it.
}
